Brings basic docs site to life with dark mode
