+++ .gitignore
@@ -35,3 +35,4 @@ package-lock.json
 /.yarn/install-state.gz
 /.yarn/cache
 /.yarn/build-state.yml
+.docusaurus
