+++ website/.gitignore
@@ -9,3 +9,4 @@ website/node_modules
 
 website/i18n/*
 !website/i18n/en.json
+.docusaurus/
