+++ website/blog/2020-05-25-7.10.0.md
@@ -38,7 +38,8 @@ Now that it has reached Stage 4, _parsing_ for [`import.meta`](https://github.co
 console.log(import.meta); // { url: "file:///home/user/my-module.js" }
 ```
 
-### Transforming `\u{...}`-style Unicode escapes ([#11377](https://github.com/babel/babel/pull/11377))
+{/* Todo: strek */}
+<!-- ### Transforming `\u{...}`-style Unicode escapes ([#11377](https://github.com/babel/babel/pull/11377)) -->
 
 We also discovered that we didn't have support for compiling a 5-year-old ECMAScript feature: `\u{...}`-style Unicode escapes! Thanks to [Justin](https://github.com/jridgewell), `@babel/preset-env` can now compile them in strings and identifiers by default.
