+++ website/sidebars.json
@@ -65,10 +65,9 @@
     ],
     "@babel/preset-env": [
       {
-        "type": "subcategory",
+        "type": "category",
         "label": "ES2022",
-        "ids": [
-          "babel-plugin-proposal-class-private-methods",
+        "items": [
           "babel-plugin-proposal-class-properties",
           "babel-plugin-proposal-class-static-block",
           "babel-plugin-proposal-private-property-in-object",
@@ -76,17 +75,17 @@
         ]
       },
       {
-        "type": "subcategory",
+        "type": "category",
         "label": "ES2021",
-        "ids": [
+        "items": [
           "babel-plugin-proposal-logical-assignment-operators",
           "babel-plugin-proposal-numeric-separator"
         ]
       },
       {
-        "type": "subcategory",
+        "type": "category",
         "label": "ES2020",
-        "ids": [
+        "items": [
           "babel-plugin-proposal-export-namespace-from",
           "babel-plugin-proposal-nullish-coalescing-operator",
           "babel-plugin-proposal-optional-chaining",
@@ -96,17 +95,17 @@
         ]
       },
       {
-        "type": "subcategory",
+        "type": "category",
         "label": "ES2019",
-        "ids": [
+        "items": [
           "babel-plugin-proposal-optional-catch-binding",
           "babel-plugin-proposal-json-strings"
         ]
       },
       {
-        "type": "subcategory",
+        "type": "category",
         "label": "ES2018",
-        "ids": [
+        "items": [
           "babel-plugin-proposal-async-generator-functions",
           "babel-plugin-proposal-object-rest-spread",
           "babel-plugin-proposal-unicode-property-regex",
@@ -115,21 +114,20 @@
         ]
       },
       {
-        "type": "subcategory",
+        "type": "category",
         "label": "ES2017",
-        "ids": ["babel-plugin-transform-async-to-generator"]
+        "items": ["babel-plugin-transform-async-to-generator"]
       },
       {
-        "type": "subcategory",
+        "type": "category",
         "label": "ES2016",
-        "ids": ["babel-plugin-transform-exponentiation-operator"]
+        "items": ["babel-plugin-transform-exponentiation-operator"]
       },
       {
-        "type": "subcategory",
+        "type": "category",
         "label": "ES2015",
-        "ids": [
+        "items": [
           "babel-plugin-transform-arrow-functions",
-          "babel-plugin-transform-block-scoped",
           "babel-plugin-transform-block-scoping",
           "babel-plugin-transform-classes",
           "babel-plugin-transform-computed-properties",
@@ -152,14 +150,14 @@
         ]
       },
       {
-        "type": "subcategory",
+        "type": "category",
         "label": "ES5",
-        "ids": ["babel-plugin-transform-property-mutators"]
+        "items": ["babel-plugin-transform-property-mutators"]
       },
       {
-        "type": "subcategory",
+        "type": "category",
         "label": "ES3",
-        "ids": [
+        "items": [
           "babel-plugin-transform-member-expression-literals",
           "babel-plugin-transform-property-literals",
           "babel-plugin-transform-reserved-words"
